tty/console: use SPCR table to define console
If console= is not added to the kernel command line, the console
is not registered until much further into the booting process. This patch
adds support to parse the SPCR ACPI table to pull console support out,
then use the appropriate drivers to set up console support earlier in the
boot process.

Signed-off-by: Jon Masters <jcm@redhat.com>
[rebased and cleaned up]
Signed-off-by: Torez Smith <torez@redhat.com>
[reworked to use _CRS, moved to drivers/acpi]
Signed-off-by: Leif Lindholm <leif.lindholm@linaro.org>