{temphack} DO NOT UPSTREAM - Fix xgene gic acpi init
Current X-Gene firmware place gic base address at 64K boundary
which breaks with the recent EOI patches. The same problem
occurred with DT but Marc Zyngier provided a patch for that.
This patch does the same thing for ACPI boots. APM needs to
either change firmware to play nice with EOI patches or get
a patch similar to this upstream.

Signed-off-by: Mark Salter <msalter@redhat.com>
Signed-off-by: Ricardo Salveti <ricardo.salveti@linaro.org>