PCI: Remove const from the pci_dev struct passed to pci_fixup_irqs
The pci_fixup_irqs function currently gets passed a const struct pci_dev *.
This is fine for the current implementation which never modifies the pci_dev
however the deferred IRQ assignment version may have to modify this pci_dev
on some architectures so the const is no longer correct.

Therefore we remove it, fortunately no users currently seem to rely on it
being a const so no changes to callers are needed.

Signed-off-by: Matthew Minter <matt@masarand.com>