PCI: Add pci_assign_irq function and have pci_fixup_irqs use it
Here we delete the static pdev_fixup_irq function which is currently
what pci_fixup_irqs uses to actually assign the irqs and replace it
with the pci_assign_irq function which changes the interface and
makes use of the new function pointers stored in the host bridge
structure.

Eventually this will allow pci_fixup_irqs to be removed entirely
and the new deffered assignment code path will call pci_assign_irq
directly. However to ensure current users continue to work a
new implementation of pci_fixup_irqs is introduced which simply
wraps the functionality of pci_assign_irq.

Signed-off-by: Matthew Minter <matt@masarand.com>