PCI: Add a call to pci_assign_irq in do_pci_enable_device
The pci_assign_irq function allows assignment of an irq to devices
during device enable time rather than only at boot. Therefore call it
in the do_pci_enable_device function during the enable device
code path so this assignment can be performed.

This patch will do nothing on arches which do not set the irq mapping
function pointers and is therefore currently a noop, however as support
for these function pointers is added to arch specific code this will
cause irq asignment to migrate to device enable time allowing the new
codepaths to be used.

Signed-off-by: Matthew Minter <matt@masarand.com>