devicetree: bindings: hisi_sas: add v2 HW bindings
Add the dt bindings for HiSi SAS controller v2 HW.

The main difference in the controller from dt perspective
is interrupts. The v2 controller does not have dedicated
fatal and broadcast interrupts - they are multiplexed on
the channel interrupt.

Each SAS v2 controller can issue upto 64 commands
(or connection requests) on the system bus without waiting
for a response - this is know as am-max-transmissions.
In hip06, sas controller #1 has a limitation that it has to
limit am-max-transmissions to 32 - this limitation is due
to chip system bus design. It is not anticipated that any
future chip incorporating v2 controller will have such a
limitation.

Signed-off-by: John Garry <john.garry@huawei.com>
Signed-off-by: Hanjun Guo <hanjun.guo@linaro.org>