SMAF: SMAF TA for secure/nonsecure memory allocation
This change introduce a SMAF allocator based on the SMAF TA.
The allocator invokes the OPTEE SMAF trusted application to
request allocation of memory. Such memory can be secured
and not.

This patch currently relies on a non-official (not yet upstreamed)
OPTEE kernel API. It can be fetched from the linaro-swg linux kernel.

Signed-off-by: Pascal Brand <pascal.brand@st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>