staging: android: ion: Add UNMAPPED heap type and helper
The "unmapped" heap is very similar to the carveout heap except
the backing memory is presumed to be unmappable by the host, in
my specific case due to firewalls. This memory can still be
allocated from and used by devices that do have access to the
backing memory.

Based originally on the secure/unmapped heap from Linaro for
the OP-TEE SDP implementation, this was re-written to match
the carveout heap helper code.

Suggested-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Andrew F. Davis <afd@ti.com>

[fix inline function in ion.h]
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>