firmware: add support for Arm Secure Partition Client Interface  (SPCI)
This patch adds support for the SPCI specification in the form of a platform
driver. The driver consumes the SPCI implementation exported by the SPM to
provide an API to clients in the OS that enables message passing with a secure
partition.

This version of the driver is based upon the SPCI Alpha2 specification. It does
not adhere to the specification religiously as the intent is to find out where
the spec falls short. Hence, there are new features and small tweaks on top of
Alpha2. Once finalised, these changes will be recorded in the next spec
revision.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
---
 drivers/firmware/arm_spci.c   | 665 ++++++++++++++++++++++++++++++++++++++++++
 include/linux/arm_spci.h      | 314 ++++++++++++++++++++
 include/linux/spci_protocol.h |  52 ++++
 3 files changed, 1031 insertions(+)
 create mode 100644 drivers/firmware/arm_spci.c
 create mode 100644 include/linux/arm_spci.h
 create mode 100644 include/linux/spci_protocol.h