tee: add SPCI specific invoke function
Arm Secure Partition Client interface specification describes SMC ABIs to
standardise communication between a OS in normal world and a secure partition in
the secure world. OP-TEE is an example of a secure partition.

By default, the OP-TEE driver uses an ABI specific to it for communication with
the OP-TEE kernel in S-EL1. This ABI uses the SMC or HVC instruction as its
conduit. The SPCI driver introduces a standard "transport" or "conduit" that the
OP-TEE driver can use for communication instead of relying on a private ABI.

This patch adds an invoke function that uses the SPCI message passing protocol
for communication. SPCI support is based on the SPCI Alpha 2 specification and
is by no means complete. For starters, the SPCI driver exports an interface that
enables the OP-TEE driver to send a message and blocking until a response is
available. This matches the semantics of the existing private ABI and the
scheduling model B in the SPCI Alpha 2 spec. A richer interface and support for
scheduling model A will be added in future.

Signed-off-by: Achin Gupta <achin.gupta@arm.com>
---
 drivers/tee/optee/core.c | 52 +++++++++++++++++++++++++++++++++++++++++++++++-
 1 file changed, 51 insertions(+), 1 deletion(-)