scmi/optee: support multile message per channel
In order to support more that 1 request at the same time per channel,
we must increase the size of the allocated shared memory and split it
in sevral different message. We map the pre allocated scmi message id
to the offset in the shared memory.

The scmi pre allocated id is saved in xfer->hdr.seq

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>